requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0 